change location of socket related operations to service
