+++ src/controllers/HelpController.js
@@ -16,16 +16,6 @@ class HelpController {
 
     try {
       const result = await this.HelpService.createHelp(data);
-      const user = await this.UserService.getUser({id: data.ownerId})
-      let help = JSON.parse(JSON.stringify(result));
-      help.user = [user]
-      const userCoords = {
-        longitude: user.location.coordinates[0],
-        latitude: user.location.coordinates[1]
-      }
-      const sendSocketMessageTo = findConnections(userCoords)
-
-      sendMessage(sendSocketMessageTo, 'new-help', help)
 
       res.status(201).json(result);
       next();
@@ -97,15 +87,6 @@ class HelpController {
 
     try {
       const result = await this.HelpService.deleteHelpLogically(id);
-      console.log(result)
-      const user = await this.UserService.getUser({id: result.ownerId})
-      const userCoords = {
-        longitude: user.location.coordinates[0],
-        latitude: user.location.coordinates[1]
-      }
-
-      const sendSocketMessageTo = findConnections(userCoords)
-      sendMessage(sendSocketMessageTo, 'delete-help', id)
 
       res.status(200).json(result);
       return next();
